backend/pyproject.toml
[tool.poetry]
name = "fitness-habit-tracker-backend"
version = "0.1.0"
description = "FastAPI backend for Fitness & Habit Tracker"
authors = ["Your Name <you@example.com>"]
license = "Apache-2.0"
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {version = "^0.30.0", extras = ["standard"]}
sqlalchemy = "^2.0.30"
alembic = "^1.13.2"
psycopg2-binary = "^2.9.9"
python-jose = {version = "^3.3.0", extras = ["cryptography"]}
passlib = {version = ">=1.7.4", extras = ["bcrypt"]}
pydantic = "^2.7.0"
pydantic-settings = "^2.2.1"
fastapi-limiter = "^0.1.6"
redis = "^5.0.4"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
ruff = "^0.4.4"
pytest = "^8.2.0"
pytest-asyncio = "^0.23.6"
httpx = "^0.27.0"
coverage = "^7.5.3"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP", "B"]
ignore = []

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


